feat(rag): migrate post vector indexing to EventBridge + rag/qdrant
- CDK: Add PostVectorIndex pipe (POST INSERT/MODIFY, status=PUBLISHED)
  and PostVectorDelete pipe (POST REMOVE, OldImage status=PUBLISHED)
- run.rs: Add PostVectorIndex and PostVectorDelete DetailType handlers
  routing to rag::qdrant::indexers::post_indexer
- Create post_indexer.rs with index_post(Post) and delete_post_index(Post)
- Create PostPayload in rag/qdrant/payloads/post_payload.rs
- Add Post variant to QdrantIndexType
- Remove inline Qdrant calls from update_post.rs and delete_post.rs
- Delete posts/services/vector_indexer.rs (migrated to rag module)
- Make posts::utils public for extract_plain_text access

+++ app/ratel/src/features/ai_moderator/controllers/upload_material.rs
@@ -27,7 +27,7 @@ pub async fn upload_material(
 
     let opt = AiModeratorMaterial::opt()
         .sk(EntityType::AiModeratorMaterial(String::default()).to_string())
-        .limit((MAX_MATERIALS + 1) as i64);
+        .limit((MAX_MATERIALS + 1) as i32);
     let (existing, _) = AiModeratorMaterial::query(cli, &pk, opt).await?;
     if existing.len() >= MAX_MATERIALS {
         return Err(AiModeratorError::MaterialLimitReached.into());
